MPI_Comm abstraction and tests.
- Removed dependency on mpi.h for all files but
  abstraction implementation.
- Created CommFactory singleton that is exposed via
  C++ interface.
- Added new exception type and message.
- Added MPI implementation unit test suite.
- MPI communitcation implementation developed as plugin.
- Environment variable added to specify desired
  communication plugin.
- Fixed test scripts that were broken by addition of
  communication plugin.

+++ src/Profile.cpp
@@ -38,7 +38,6 @@
 #include <sys/sysctl.h>
 #endif
 
-#include <algorithm>
 #include <iostream>
 #include <sstream>
 
@@ -50,25 +49,22 @@
 #include "geopm.h"
 #include "geopm_ctl.h"
 #include "geopm_message.h"
-#include "geopm_time.h"
 #include "geopm_signal_handler.h"
 #include "geopm_sched.h"
 #include "Profile.hpp"
+#include "ProfileSampler.hpp"
 #include "ProfileThread.hpp"
 #include "Exception.hpp"
 #include "geopm_env.h"
 #include "ProfileTable.hpp"
 #include "geopm_sched.h"
 #include "config.h"
-
-static bool geopm_prof_compare(const std::pair<uint64_t, struct geopm_prof_message_s> &aa, const std::pair<uint64_t, struct geopm_prof_message_s> &bb)
-{
-    return geopm_time_comp(&(aa.second.timestamp), &(bb.second.timestamp));
-}
+#include "Comm.hpp"
 
 static geopm::Profile &geopm_default_prof(void)
 {
-    static geopm::Profile default_prof(geopm_env_profile(), MPI_COMM_WORLD);
+    const geopm::IComm *tmp_comm = geopm::geopm_get_comm("MPIComm");
+    static geopm::Profile default_prof(geopm_env_profile(), tmp_comm);
     return default_prof;
 }
 
@@ -208,8 +204,7 @@ extern "C"
 
 namespace geopm
 {
-
-    Profile::Profile(const std::string prof_name, MPI_Comm comm)
+    Profile::Profile(const std::string prof_name, const IComm *comm)
         : m_is_enabled(true)
         , m_prof_name(prof_name)
         , m_curr_region_id(0)
@@ -225,7 +220,7 @@ namespace geopm
         , m_tprof_table(NULL)
         , M_OVERHEAD_FRAC(0.01)
         , m_scheduler(NULL)
-        , m_shm_comm(MPI_COMM_NULL)
+        , m_shm_comm(NULL)
         , m_rank(0)
         , m_shm_rank(0)
         , m_is_first_sync(true)
@@ -243,11 +238,11 @@ namespace geopm
         int shm_num_rank = 0;
 
         m_scheduler = new SampleScheduler(M_OVERHEAD_FRAC);
-        MPI_Comm_rank(comm, &m_rank);
-        geopm_comm_split_shared(comm, "prof", &m_shm_comm);
-        PMPI_Comm_rank(m_shm_comm, &m_shm_rank);
-        PMPI_Comm_size(m_shm_comm, &shm_num_rank);
-        PMPI_Barrier(m_shm_comm);
+        m_rank = comm->rank();
+        m_shm_comm = comm->split("prof", IComm::M_COMM_SPLIT_TYPE_SHARED);
+        m_shm_rank = m_shm_comm->rank();
+        shm_num_rank = m_shm_comm->num_rank();
+        m_shm_comm->barrier();
 
         std::string key(geopm_env_shmkey());
         key += "-sample";
@@ -264,7 +259,7 @@ namespace geopm
             m_is_enabled = false;
             return;
         }
-        PMPI_Barrier(m_shm_comm);
+        m_shm_comm->barrier();
         if (!m_shm_rank) {
             m_ctl_shmem->unlink();
         }
@@ -273,7 +268,7 @@ namespace geopm
 
         init_cpu_list();
 
-        PMPI_Barrier(m_shm_comm);
+        m_shm_comm->barrier();
         m_ctl_msg->step();
         m_ctl_msg->wait();
 
@@ -298,7 +293,7 @@ namespace geopm
                     }
                 }
             }
-            PMPI_Barrier(m_shm_comm);
+            m_shm_comm->barrier();
         }
 
         if (!m_shm_rank) {
@@ -316,28 +311,32 @@ namespace geopm
                 }
             }
         }
-        PMPI_Barrier(m_shm_comm);
+        m_shm_comm->barrier();
         m_ctl_msg->step();
         m_ctl_msg->wait();
+
+        std::ostringstream table_shm_key;
+        table_shm_key << key <<  "-"  << m_rank;
+        m_table_shmem = new SharedMemoryUser(table_shm_key.str(), 3.0);
+        m_shm_comm->barrier();
+        if (!m_shm_rank) {
+            m_table_shmem->unlink();
+        }
+        m_shm_comm->barrier();
+
+        m_table_buffer = m_table_shmem->pointer();
+        m_table = new ProfileTable(m_table_shmem->size(), m_table_buffer);
 
         std::string tprof_key(geopm_env_shmkey());
         tprof_key += "-tprof";
         m_tprof_shmem = new SharedMemoryUser(tprof_key, 3.0);
-        PMPI_Barrier(m_shm_comm);
+        m_shm_comm->barrier();
         if (!m_shm_rank) {
             m_tprof_shmem->unlink();
         }
         m_tprof_table = new ProfileThreadTable(m_tprof_shmem->size(), m_tprof_shmem->pointer());
 
-        std::ostringstream table_shm_key;
-        table_shm_key << key <<  "-"  << m_rank;
-        m_table_shmem = new SharedMemoryUser(table_shm_key.str(), 3.0);
-        m_table_shmem->unlink();
-
-        m_table_buffer = m_table_shmem->pointer();
-        m_table = new ProfileTable(m_table_shmem->size(), m_table_buffer);
-
-        PMPI_Barrier(m_shm_comm);
+        m_shm_comm->barrier();
         m_ctl_msg->step();
         m_ctl_msg->wait();
         geopm_pmpi_prof_enable(1);
@@ -371,7 +370,7 @@ namespace geopm
 #endif
 
         geopm_pmpi_prof_enable(0);
-        PMPI_Barrier(m_shm_comm);
+        m_shm_comm->barrier();
         m_ctl_msg->step();
         m_ctl_msg->wait();
 
@@ -384,7 +383,7 @@ namespace geopm
         if (geopm_env_report_verbosity()) {
             print(geopm_env_report(), geopm_env_report_verbosity());
         }
-        PMPI_Barrier(m_shm_comm);
+        m_shm_comm->barrier();
         m_ctl_msg->step();
         m_is_enabled = false;
     }
@@ -428,7 +427,7 @@ namespace geopm
         if (!m_curr_region_id && region_id) {
             if (!geopm_region_id_is_mpi(region_id) &&
                 geopm_env_do_region_barrier()) {
-                PMPI_Barrier(m_shm_comm);
+                m_shm_comm->barrier();
             }
             m_curr_region_id = region_id;
             m_num_enter = 0;
@@ -491,7 +490,7 @@ namespace geopm
         if (!m_num_enter) {
             if (!geopm_region_id_is_mpi(region_id) &&
                 geopm_env_do_region_barrier()) {
-                PMPI_Barrier(m_shm_comm);
+                m_shm_comm->barrier();
             }
             if (geopm_region_id_is_mpi(region_id)) {
                 m_curr_region_id = geopm_region_id_set_mpi(m_parent_region);
@@ -557,7 +556,7 @@ namespace geopm
 #endif
 
         struct geopm_prof_message_s sample;
-        PMPI_Barrier(m_shm_comm);
+        m_shm_comm->barrier();
         if (!m_shm_rank) {
             sample.rank = m_rank;
             sample.region_id = GEOPM_REGION_ID_EPOCH;
@@ -614,7 +613,7 @@ namespace geopm
         int is_done = 0;
         int is_all_done = 0;
 
-        PMPI_Barrier(m_shm_comm);
+        m_shm_comm->barrier();
         m_ctl_msg->step();
         m_ctl_msg->wait();
 
@@ -633,17 +632,17 @@ namespace geopm
         strncpy(buffer_ptr, m_prof_name.c_str(), buffer_remain - 1);
         buffer_offset += m_prof_name.length() + 1;
         while (!is_all_done) {
-            PMPI_Barrier(m_shm_comm);
+            m_shm_comm->barrier();
             m_ctl_msg->loop_begin();
 
             is_done = m_table->name_fill(buffer_offset);
-            PMPI_Allreduce(&is_done, &is_all_done, 1, MPI_INT, MPI_LAND, m_shm_comm);
+            is_all_done = m_shm_comm->test(is_done);
 
             m_ctl_msg->step();
             m_ctl_msg->wait();
             buffer_offset = 0;
         }
-        PMPI_Barrier(m_shm_comm);
+        m_shm_comm->barrier();
         m_ctl_msg->step();
         m_ctl_msg->wait();
 
@@ -655,8 +654,8 @@ namespace geopm
                                      m_overhead_time,
                                      m_overhead_time_shutdown};
         double max_overhead[3] = {};
-        MPI_Reduce(overhead_buffer, max_overhead, 3,
-                   MPI_DOUBLE, MPI_MAX, 0, MPI_COMM_WORLD);
+        const IComm *tmp_comm = geopm::geopm_get_comm("MPIComm");
+        tmp_comm->reduce_max(overhead_buffer, max_overhead, 3, 0);
         if (!m_rank) {
             std::cout << "GEOPM startup (seconds):  " << max_overhead[0] << std::endl;
             std::cout << "GEOPM runtime (seconds):  " << max_overhead[1] << std::endl;
@@ -691,227 +690,5 @@ namespace geopm
     IProfileThreadTable *Profile::tprof_table(void)
     {
         return m_tprof_table;
-    }
-
-
-    const struct geopm_prof_message_s GEOPM_INVALID_PROF_MSG = {-1, 0, {{0, 0}}, -1.0};
-
-    ProfileSampler::ProfileSampler(size_t table_size)
-        : m_table_size(table_size)
-        , m_do_report(false)
-        , m_tprof_shmem(NULL)
-        , m_tprof_table(NULL)
-    {
-        std::string sample_key(geopm_env_shmkey());
-        sample_key += "-sample";
-        m_ctl_shmem = new SharedMemory(sample_key, table_size);
-        m_ctl_msg = new ControlMessage((struct geopm_ctl_message_s *)m_ctl_shmem->pointer(), true, true);
-
-        std::string tprof_key(geopm_env_shmkey());
-        tprof_key += "-tprof";
-        size_t tprof_size = 64 * geopm_sched_num_cpu();
-        m_tprof_shmem = new SharedMemory(tprof_key, tprof_size);
-        m_tprof_table = new ProfileThreadTable(tprof_size, m_tprof_shmem->pointer());
-    }
-
-    ProfileSampler::~ProfileSampler(void)
-    {
-        delete m_tprof_table;
-        delete m_tprof_shmem;
-        for (auto it = m_rank_sampler.begin(); it != m_rank_sampler.end(); ++it) {
-            delete (*it);
-        }
-        delete m_ctl_shmem;
-    }
-
-    void ProfileSampler::initialize(int &rank_per_node)
-    {
-        std::ostringstream shm_key;
-
-        m_ctl_msg->wait();
-        m_ctl_msg->step();
-        m_ctl_msg->wait();
-
-        std::set<int> rank_set;
-        for (int i = 0; i < GEOPM_MAX_NUM_CPU; i++) {
-            if (m_ctl_msg->cpu_rank(i) >= 0) {
-                (void) rank_set.insert(m_ctl_msg->cpu_rank(i));
-            }
-        }
-
-        for (auto it = rank_set.begin(); it != rank_set.end(); ++it) {
-            shm_key.str("");
-            shm_key << m_ctl_shmem->key() <<  "-"  << *it;
-            m_rank_sampler.push_front(new ProfileRankSampler(shm_key.str(), m_table_size));
-        }
-        rank_per_node = rank_set.size();
-        m_ctl_msg->step();
-        m_ctl_msg->wait();
-        m_ctl_msg->step();
-    }
-
-    void ProfileSampler::cpu_rank(std::vector<int> &cpu_rank)
-    {
-        uint32_t num_cpu = geopm_sched_num_cpu();
-        cpu_rank.resize(num_cpu);
-        if (num_cpu > GEOPM_MAX_NUM_CPU) {
-            throw Exception("ProfileSampler::cpu_rank: Number of online CPUs is greater than GEOPM_MAX_NUM_CPU", GEOPM_ERROR_RUNTIME, __FILE__, __LINE__);
-        }
-
-        for (unsigned cpu = 0; cpu < num_cpu; ++cpu) {
-            cpu_rank[cpu] = m_ctl_msg->cpu_rank(cpu);
-        }
-    }
-
-    size_t ProfileSampler::capacity(void)
-    {
-        size_t result = 0;
-        for (auto it = m_rank_sampler.begin(); it != m_rank_sampler.end(); ++it) {
-            result += (*it)->capacity();
-        }
-        return result;
-    }
-
-    void ProfileSampler::sample(std::vector<std::pair<uint64_t, struct geopm_prof_message_s> > &content, size_t &length, MPI_Comm comm)
-    {
-        length = 0;
-        if (m_ctl_msg->is_sample_begin() ||
-            m_ctl_msg->is_sample_end()) {
-            auto content_it = content.begin();
-            for (auto rank_sampler_it = m_rank_sampler.begin();
-                 rank_sampler_it != m_rank_sampler.end();
-                 ++rank_sampler_it) {
-                size_t rank_length = 0;
-                (*rank_sampler_it)->sample(content_it, rank_length);
-                content_it += rank_length;
-                length += rank_length;
-            }
-            if (m_ctl_msg->is_sample_end()) {
-                PMPI_Barrier(comm);
-                m_ctl_msg->step();
-                while (!m_ctl_msg->is_name_begin() &&
-                       !m_ctl_msg->is_shutdown()) {
-                    geopm_signal_handler_check();
-                }
-                if (m_ctl_msg->is_name_begin()) {
-                    region_names();
-                }
-            }
-        }
-        else if (!m_ctl_msg->is_shutdown()) {
-            throw Exception("ProfileSampler: invalid application status, expected shutdown status", GEOPM_ERROR_RUNTIME, __FILE__, __LINE__);
-        }
-    }
-
-    bool ProfileSampler::do_shutdown(void)
-    {
-        return m_ctl_msg->is_shutdown();
-    }
-
-    bool ProfileSampler::do_report(void)
-    {
-        return m_do_report;
-    }
-
-    void ProfileSampler::region_names(void)
-    {
-        m_ctl_msg->step();
-
-        bool is_all_done = false;
-        while (!is_all_done) {
-            m_ctl_msg->loop_begin();
-            m_ctl_msg->wait();
-            is_all_done = true;
-            for (auto it = m_rank_sampler.begin(); it != m_rank_sampler.end(); ++it) {
-                if (!(*it)->name_fill(m_name_set)) {
-                    is_all_done = false;
-                }
-            }
-            m_ctl_msg->step();
-            if (!is_all_done && m_ctl_msg->is_shutdown()) {
-                throw Exception("ProfileSampler::region_names(): Application shutdown while report was being generated", GEOPM_ERROR_RUNTIME, __FILE__, __LINE__);
-            }
-        }
-        m_rank_sampler.front()->report_name(m_report_name);
-        m_rank_sampler.front()->profile_name(m_profile_name);
-        m_do_report = true;
-
-        m_ctl_msg->wait();
-        m_ctl_msg->step();
-        m_ctl_msg->wait();
-    }
-
-    void ProfileSampler::name_set(std::set<std::string> &region_name)
-    {
-        region_name = m_name_set;
-    }
-
-    void ProfileSampler::report_name(std::string &report_str)
-    {
-        report_str = m_report_name;
-    }
-
-    void ProfileSampler::profile_name(std::string &prof_str)
-    {
-        prof_str = m_profile_name;
-    }
-
-    IProfileThreadTable *ProfileSampler::tprof_table(void)
-    {
-        return m_tprof_table;
-    }
-
-    ProfileRankSampler::ProfileRankSampler(const std::string shm_key, size_t table_size)
-        : m_table_shmem(NULL)
-        , m_table(NULL)
-        , m_region_entry(GEOPM_INVALID_PROF_MSG)
-        , m_is_name_finished(false)
-    {
-        m_table_shmem = new SharedMemory(shm_key, table_size);
-        m_table = new ProfileTable(m_table_shmem->size(), m_table_shmem->pointer());
-    }
-
-    ProfileRankSampler::~ProfileRankSampler()
-    {
-        delete m_table;
-        delete m_table_shmem;
-    }
-
-    size_t ProfileRankSampler::capacity(void)
-    {
-        return m_table->capacity();
-    }
-
-    void ProfileRankSampler::sample(std::vector<std::pair<uint64_t, struct geopm_prof_message_s> >::iterator content_begin, size_t &length)
-    {
-        m_table->dump(content_begin, length);
-        std::stable_sort(content_begin, content_begin + length, geopm_prof_compare);
-    }
-
-    bool ProfileRankSampler::name_fill(std::set<std::string> &name_set)
-    {
-        size_t header_offset = 0;
-
-        if (!m_is_name_finished) {
-            if (name_set.empty()) {
-                m_report_name = (char *)m_table_shmem->pointer();
-                header_offset += m_report_name.length() + 1;
-                m_prof_name = (char *)m_table_shmem->pointer() + header_offset;
-                header_offset += m_prof_name.length() + 1;
-            }
-            m_is_name_finished = m_table->name_set(header_offset, name_set);
-        }
-
-        return m_is_name_finished;
-    }
-
-    void ProfileRankSampler::report_name(std::string &report_str)
-    {
-        report_str = m_report_name;
-    }
-
-    void ProfileRankSampler::profile_name(std::string &prof_str)
-    {
-        prof_str = m_prof_name;
     }
 }
